requirements.txt
click==8.1.7
distro==1.9.0
fastapi==0.95.2
gunicorn==20.1.0
httptools==0.6.1
icu==0.0.1
mailjet-rest==1.3.4
openai==1.30.1
polyglot==16.7.4
pycld2==0.41
pydantic==1.10.15
PyICU==2.13.1
python-dotenv==1.0.1
starlette==0.27.0
tqdm==4.66.4
uvicorn==0.22.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
xlwings==0.30.1
